requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.4.2
python-multipart==0.0.6
cryptography==41.0.5
bcrypt==4.0.1
python-jose==3.3.0
passlib==1.7.4 